+++ environment.sample.sh
@@ -1,5 +1,8 @@
 export RACKSPACE_USERNAME=
 export RACKSPACE_APIKEY=
 export RACKSPACE_REGION=
-export RACKSPACE_CONTAINER=
+export RACKSPACE_SERVICENET=
+export ADMIN_APIKEY=
+export CONTENT_CONTAINER=
+export ASSET_CONTAINER=
 export CONTENT_LOG_LEVEL=
